fix[api]: route video generation cancellation through request registry
